feat: refactor AsyncAPI version handling and improve path resolution logic
