+++ asyncapi-front-end/src/main/java/org/apiaddicts/apitools/dosonarapi/api/AsyncApiVersion.java
@@ -17,7 +17,7 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package org.apiaddicts.apitools.dosonarapi.plugin;
+package org.apiaddicts.apitools.dosonarapi.api;
 
 public enum AsyncApiVersion {
   v2_x("2.x"),
